add pin prompt screen and biometric prompt

+++ app/src/main/java/id/dev/spendless/core/domain/model/UserSecurity.kt
@@ -2,6 +2,6 @@ package id.dev.spendless.core.domain.model
 
 data class UserSecurity(
     val biometricPromptEnable: Boolean,
-    val sessionExpiryDuration: Int,
-    val lockedOutDuration: Int,
+    val sessionExpiryDuration: Long,
+    val lockedOutDuration: Long,
 )
